requirements.txt
# Jalankan perintah dibawah
# python -m venv venv
# venv\Scripts\activate
# pip install -r requirements.txt

discord.py==2.5.2
google-generativeai==0.8.5
mplfinance==0.12.10b0
numpy==2.2.6
openai==1.82.0
pandas==2.2.3
python-binance==1.0.29
python-dotenv==1.1.0
yfinance==0.2.61
